Import/export update plus rename modules to exported names
