+++ src/plugins/index.js
@@ -1,2 +1,2 @@
-module.exports.VersioningPlugin = require('./viewdb_versioning_plugin');
-module.exports.TimestampPlugin = require('./viewdb_timestamp_plugin');
+export { default as TimestampPlugin } from './ViewDBTimestampPlugin';
+export { default as VersioningPlugin } from './ViewDBVersioningPlugin';
